[Update] 更新发布脚本

+++ build.ps1
@@ -98,8 +98,12 @@ Copy-Item $ADMIN_APP_DIST_DIR $ADMIN_APP_DIR -Recurse
 Set-Location $CURRENT_DIR
 . ./base.ps1
 
-Write-Output "EMPTY Publish Directory"
-Remove-Item -LiteralPath $PUBLISHED_DIR -Force -Recurse
+$HAS_PUBLISH_DIR = Test-Path $PUBLISHED_DIR
+
+if ($HAS_PUBLISH_DIR) {
+    Write-Output "EMPTY Publish Directory"
+    Remove-Item -LiteralPath $PUBLISHED_DIR -Force -Recurse
+}
 
 # 发布站点到发布目录
 $BuildWebCommand = "dotnet publish {0} -c Release -o {1}" -f $WEB_DIR, $PUBLISHED_DIR
